pyproject.toml
[tool.poetry]
name = "lamochka-llm-council"
version = "1.0.0"
description = "Multi-LLM Council for collaborative AI decision-making with role-based debates"
authors = ["Ruslan Lama <ruslan@example.com>"]
readme = "README.md"
license = "MIT"
keywords = ["llm", "ai", "council", "multi-agent", "debate", "grok", "claude", "gemini", "chatgpt"]
repository = "https://github.com/llamich/lamochka-llm-council"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
httpx = "^0.26.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
openai = "^1.10.0"
anthropic = "^0.18.0"
google-generativeai = "^0.4.0"
rich = "^13.7.0"
PyYAML = "^6.0.1"
tenacity = "^8.2.3"
duckduckgo-search = "^4.4.0"
aiohttp = "^3.9.0"
ollama = "^0.1.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.2.0"

[tool.poetry.scripts]
council = "main:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
